catalogue/requirements3.txt
grpcio==1.51.3
grpcio-tools==1.30.0
googleapis-common-protos==1.58.0
protobuf==3.19.5
Jinja2
